+++ kwsCheckVariablePerLine.cxx
@@ -117,7 +117,7 @@ bool Parser::CheckVariablePerLine(unsigned long max)
                 }
               // Check if we are not at the end of the line
               bool endofline = true;
-              int eof = pos+1;
+              long eof = pos+1;
               while(eof < (int)line.size())
                 {
                 if(line[eof] != ' ' && line[eof] != '\n'
